Moved weather logic to WeatherService
